requirements.txt
hydra-core
tqdm
wandb
gpustat
loguru
einops==0.7.0
fancy-einsum==0.0.3
huggingface-hub==0.21.1
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
omegaconf==2.3.0
safetensors==0.4.2
tokenizers==0.15.2
huggingface_hub
numpy
evaluate
mauve-text
datasets==2.18.0
fsspec==2024.2.0
lightning==2.2.1
pandas==2.2.1
timm==0.9.16
transformers==4.38.2
wandb==0.13.5
jupyter
pyyaml
ninja
vendi_score[all]
